+++ admin/delete-user.php
@@ -1,5 +1,12 @@
 <?php
-require_once '../config/config.php';
+require_once '../config/autoload.php';
+if (!isUserLoggedIn() && !isUserLoggedInAdmin()) {
+    return redirect('/users/login.php');
+}
+/***
+ * @var $db
+ */
+
 $id = $_GET['id'];
 $query = mysqli_query($db,"DELETE FROM users WHERE id=$id");
 if($query){
